Adding cast so loss percent is calculated correctly.
(Ok, ok Eric. You were right, I should have held off longer before making another public
release.)

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@899 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/stats.c
@@ -1451,12 +1451,12 @@ OWPStatsPrintSummary(
      * lost % is 0 if sent == 0.
      */
     if(stats->sent > 0){
-        d1 = stats->lost/stats->sent;
+        d1 = (double)stats->lost/stats->sent;
     }
     else{
         d1 = 0.0;
     }
-    fprintf(output,"%u sent, %u lost (%.1f%%), %u duplicates\n",
+    fprintf(output,"%u sent, %u lost (%.3f%%), %u duplicates\n",
             stats->sent,stats->lost,100.0*d1,stats->dups);
 
     /*
